[USER NAVIGATION] adapting pages to match new UX page navigation home - search - login
