+++ reactApp/src/App.js
@@ -2,10 +2,11 @@ import { Routes, Route } from "react-router-dom";
 // Pages
 import Home from "pages/Home";
 import About from "pages/About";
+import Search from "pages/Search";
 import Login from "components/authentication/Login";
 import SignUp from "components/authentication/SignUp";
 import RemedyDetails from "pages/RemedyDetails";
-
+import BarNavBottom from "components/layouts/BarNavBottom";
 // Styles
 import "assets/App.css";
 import theme from "assets/theme";
@@ -18,6 +19,7 @@ function App() {
         <Routes>
           <Route exact path="/" element={<Home />}></Route>
           <Route exact path="/about" element={<About />}></Route>
+          <Route exact path="/search" element={<Search />}></Route>
           <Route exact path="/login" element={<Login />}></Route>
           <Route exact path="/signup" element={<SignUp />}></Route>
           <Route exact path="/remedies/:id" element={<RemedyDetails />}></Route>
